pyproject.toml
[tool.uv.sources]
kairix = { workspace = true }

[project]
name = "kairix-mind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gradio>=5.33.0",
    "neo4j>=5.28.1",
    "neomodel>=5.5.0",
    "python-dotenv>=1.1.0",
    "uv-build>=0.7.11",
]


tool.uv.package = true

[build-system]
requires = ["uv_build>=0.7.11,<0.8.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "kairix"

[project.scripts]
mem-ui = "kairix.ui.memory_pipeline:main"

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "ruff>=0.11.12",
]
